bad fixes just to build on 5.3

+++ src/allocator.rs
@@ -10,7 +10,7 @@ unsafe impl GlobalAlloc for KernelAllocator {
     unsafe fn alloc(&self, layout: Layout) -> *mut u8 {
         // krealloc is used instead of kmalloc because kmalloc is an inline function and can't be
         // bound to as a result
-        bindings::krealloc(ptr::null(), layout.size(), bindings::GFP_KERNEL) as *mut u8
+        bindings::krealloc(ptr::null(), layout.size() as u64, bindings::GFP_KERNEL) as *mut u8
     }
 
     unsafe fn dealloc(&self, ptr: *mut u8, _layout: Layout) {
